week_7_project/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
datetime
opendatasets
